+++ src/org/bds/test/unit/TestCasesSwitch.java
@@ -52,5 +52,45 @@ public class TestCasesSwitch extends TestCasesBase {
         // Switch: Empty statement error
         compileErrors(dir + "test68.bds", "Empty switch statement");
     }
+    @Test
+    public void test145_switch() {
+        // Switch statement: Case
+        runAndCheck(dir + "run_145.bds", "out", 3);
+    }
 
+    @Test
+    public void test146_switch_fallthrough() {
+        // Switch statement: case fallthrough
+        runAndCheck(dir + "run_146.bds", "out", 35);
+    }
+
+    @Test
+    public void test147_switch_default() {
+        // Switch statement: default
+        runAndCheck(dir + "run_147.bds", "out", 100);
+    }
+
+    @Test
+    public void test148_switch_default_fallthrough() {
+        // Switch statement: default fallthrough
+        runAndCheck(dir + "run_148.bds", "out", 700);
+    }
+
+    @Test
+    public void test153_caseInt() {
+        // Switch: Case using int
+        runAndCheck(dir + "run_153.bds", "r", "The answer");
+    }
+
+    @Test
+    public void test154_caseReal() {
+        // Switch: Case using real
+        runAndCheck(dir + "run_154.bds", "res", "OK");
+    }
+
+    @Test
+    public void test166_switch_case_return() {
+        // Switch: case + return statement
+        runAndCheck(dir + "run_166.bds", "res", "1");
+    }
 }
